scraper_ultime/requirements.txt
requests
beautifulsoup4
selenium
tqdm
PySide6
pandas
openpyxl
markdownify
pyyaml
